moving multihost.replica and client variables
Removing these attributes from the default multihost fixture
in qe_class.py to clean up the fixture.  Default is to use the
lists for replicas and clients.

Adding convenience variables to functional_services/conftest.py
in setup_session.  This is to support single replica and client
variables just for functional_services suite.

Change-Id: Idc5ff9f16b0ce4208a9586887e263432b80bd2bd
Signed-off-by: Scott Poore <spoore@redhat.com>