cert-profile lib output check updates
Changed the output checks in _run helper functions from
separate checks of stdout and stderr to check both combined at the same time.

There was a problem with some failures where expected output was seen in stdout.
Then it would check the same for stderr if stderr existed.  In cases where it
existed, it likely will have a different message and thus will not match
expected output if that was expected in stdout.  So, I combined stdout and
stderr and checked at the same time to simplify the check.

Change-Id: Ie66a0d0536af9f1be16f7f12af3dbe2408c1a969
Signed-off-by: Scott Poore <spoore@redhat.com>