qe_install domain-level should use domain_level
In the check that domain_level is a valid number in range(1, 10), the
domain-level arg is set to 0 by default.  We should instead be setting
it to the value of domain_level.

At this time, I don't know that we want to change the default from 1 to
0.  So, I'm leaving that alone.  If we need to force domain_level 0,
this change will support that by adding it to mh_cfg.yaml I think.

Change-Id: I55294eeeabaf93e39f0bafac76399b94a3962c58
Signed-off-by: Scott Poore <spoore@redhat.com>