Add create-task-repo playbook to all pytests metadata files
This commit adds support for odcs and brewtask for all pytests test suites
by including create-task-repo playbook. This playbook is needed for Gating
pipelines where we need to run the tests for a particular module-stream/rpm
version. By including this playbook, an additional repo pointing to the
module or rpm build will be installed in the SUTs. The repo will be installed
only when CI_MESSAGE is populated from Listener pipelines.