Add Resource Ready Mechanism
The purpose of this commit is to save QEs/Devel time in debugging failures
that are known about before the job starts: infrastructure outages. This
commit implements a check in all the production pipelines that is using
resources that might fail, causing precious time to be wasted figuring
out why.

Our pipelines depend on OpenShift and OpenStack resources so this
implementation is validating that these type of resources are available
before starting the execution:

- openstack.psi.redhat.com
- paas.psi.redhat.com