Remove tier2 gating pipelines
We are supposed to add tier2 gating pipelines, but it is not clear what
are the test suites that we need to run. They must be defined here:

https://gitlab.cee.redhat.com/identity-management/idm-ci/blob/master/doc/test-suite-details.yaml

This commit simply removes tier2 gating pipelines, as they are not in use and
they don't define any content that makes sense.