sudo-ipa-schema: fix expected output
When the sudo command is run on an environment without tty, and a
password is required, the command prints an error message explaining
the issue. This error message has been clarified in sudo 1.8.28
(see issue 828 https://bugzilla.sudo.ws/show_bug.cgi?id=828), meaning
a new message is used in RHEL 8.2.

The test code needs to be updated with the new expected message.

https://projects.engineering.redhat.com/browse/FREEIPA-4462