Fix QEW
1. Nightly image fix -

With the current flow, the metadata is updated correctly with the fork_updater script.
However, the metadata path is from the current repo which gets overwritten at clone step in the init phase.
Hence changing the logic to move the script after modification to /tmp
and te to use a copied version of metadata instead of one from the repo path.

2. Artifact upload fix in case of job failure -

Currently the upload-artifact.sh script is run in the script section.
If a job fails with any failure further steps in the script section are not run.
Hence, separating the logic in before_script, script, and after_script.

before_script:
contains common steps to clone and install idm-ci, modify the metadata with fork_updater script. Move the modified metadata to /tmp

script:
execute te up to test phase

after_script:
execute te --phase teardown
execute upload_artifact.sh script

Signed-off-by: bbhavsar <bbhavsar@redhat.com>